refactor node groups
